Raspberry Pi & Gamepad Patch
Two patches:
1. Ensures PyGames is at least Version 2.0.0 (and offers to upgrade). Particularly important with piOS for the Raspberry pi.
2. Tests if the gamepad the right stick is stuck down, and assumes the controller layout is different. 
In pygame, some gamepads use Axis 3 as a the X axis of the right stick, but others use it as the Left trigger. The early detection catches a potential runtime error and can save a session. A better solution for the user is the custom ini file, to confirm which controller type they have.

Warning: holding the right stick down will automatically trigger the fail-safe.